refactor handler/router
