+++ app/student/handler.js
@@ -1,95 +1,40 @@
-// uses controller(s) method(s) to compose data
+// entity based handler
+module.exports = (studentController) => ({
+  list: async (req, res) => {
+    const students = await studentController.list()
+    return res.json({ data: students })
+  },
 
-const StudentController = require('./controller')
+  create: async (req, res) => {
+    const { name, age, classId } = req.body
+    const newStudent = { name, age, classId }
 
-/**
- * List all students
- *
- */
-const list = async () => {
-  const students = await StudentController.list()
-  return students
-}
-
-/**
- * Get a specific student by their _id
- *
- */
-const getById = async id => {
-  try {
-    // check for falsy values
-    if (!id) {
-      throw new Error('id must have a value')
+    try {
+      const newlyCreatedStudent = await studentController.create(newStudent)
+      return res.json({ data: newlyCreatedStudent })
+    } catch (error) {
+      return res.status(400).json({ status: 'ERROR', error })
     }
+  },
 
-    const student = await StudentController.getById(id)
-    return student
-  } catch(error) {
-    console.error(error)
-    throw error
+  get: async (req, res) => {
+    const { id } = req.params
+
+    const student = await studentController.getById(id)
+    return res.json({ data: student })
+  },
+
+  getByClassId: async (req, res) => {
+    const { classId } = req.params
+
+    const students = await studentController.listByClass(classId)
+    return res.json({ data: students })
+  },
+
+  getByStudentName: async (req, res) => {
+    const { name } = req.params
+
+    const students = await studentController.listByName(name)
+    return res.json({ data: students })
   }
-}
-
-/**
- * List students by name
- *
- */
-const listByName = async name => {
-  try {
-    // check for falsy values
-    if (!name) {
-      throw new Error('name must have a value')
-    }
-
-    const students = await StudentController.listByName(id)
-    return students
-  } catch(error) {
-    console.error(error)
-    throw error
-  }
-}
-
-/**
- * List students in a class
- *
- */
-const listByClass = async classId => {
-  try {
-    // check for falsy values
-    if (!classId) {
-      throw new Error('class must have a value')
-    }
-
-    const students = await StudentController.listByClass(classId)
-    return students
-  } catch(error) {
-    console.error(error)
-    throw error
-  }
-}
-
-/**
- * Create a student
- *
- */
-const create = async (studentData) => {
-  try {
-    if (!studentData) {
-      throw new Error('Student data connot be null')
-    }
-
-    const newlyCreatedStudent = await StudentController.create(studentData)
-    return newlyCreatedStudent
-  } catch(error) {
-    console.error(error)
-    throw error
-  }
-}
-
-module.exports = {
-  list,
-  getById,
-  listByName,
-  listByClass,
-  create,
-}
+})
